object_detection_text/requirements.txt
opencv-python==4.9.0.80
pillow==10.2.0
requests==2.31.0
streamlit==1.30.0
ultralytics==8.1.6
